requirements.txt
requests
arpy
rpmfile
python-debian
zstandard